+++ src/data.table.h
@@ -105,3 +105,6 @@ double iquickselect(int *x, int n, int k);
 
 // openmp-utils.c
 int getDTthreads();
+void avoid_openmp_hang_within_fork();
+
+