+++ src/openmp-utils.c
@@ -1,4 +1,7 @@
 #include "data.table.h"
+#ifdef _OPENMP
+#include <pthread.h>
+#endif
 
 /* GOALS:
 * 1) By default use all CPU for end-user convenience in most usage scenarios.
@@ -9,11 +12,25 @@
 *    iii) Provide way to restrict data.table only independently of base R and 
 *         other packages using openMP
 * 4) Avoid user needing to remember to unset this control after their use of data.table
+* 5) Automatically drop down to 1 thread when called from parallel package (e.g. mclapply) to
+*    avoid the deadlock/hang (#1745 and #1727) and return to prior state afterwards.
 */
 
-// 0 means use omp_get_max_threads() default, otherwise use this many threads
-// never read directly, hence static. Go via getDTthreads()
 static int DTthreads = 0;
+// Never read directly, hence static. Always go via getDTthreads() and check that in
+// future using grep's in CRAN_Release.cmd
+
+int getDTthreads() {
+#ifdef _OPENMP
+    return DTthreads == 0 ? omp_get_max_threads() : DTthreads;  
+#else
+    return 1;
+#endif
+}
+
+SEXP getDTthreads_R() {
+    return ScalarInteger(getDTthreads());
+}
 
 SEXP setDTthreads(SEXP threads) {
     if (!isInteger(threads) || length(threads) != 1 || INTEGER(threads)[0] < 0) {
@@ -27,15 +44,20 @@ SEXP setDTthreads(SEXP threads) {
     return (R_NilValue);
 }
 
-int getDTthreads() {
+// auto avoid deadlock when data.table called from parallel::mclapply
+static int preFork_DTthreads = 0;
+void when_fork() {
+    preFork_DTthreads = DTthreads;
+    DTthreads = 1;
+}
+void when_fork_end() {
+    DTthreads = preFork_DTthreads;
+}
+void avoid_openmp_hang_within_fork() {
+    // Called once on loading data.table from init.c
 #ifdef _OPENMP
-    return DTthreads == 0 ? omp_get_max_threads() : DTthreads;  
-#else
-    return 1;
+    pthread_atfork(&when_fork, &when_fork_end, NULL);
 #endif
 }
 
-SEXP getDTthreads_R() {
-    return ScalarInteger(getDTthreads());
-}